removed unused imports, moved single test
